requirements.txt
mysql-connector-python==8.0.24
requests==2.24.0
urllib3==1.25.11